fix: audio clips missing their waveform pyramid ref render mono
A clip whose per-clip sourceAnalysisRefs is empty (created/rebuilt before the
source finished analysing, and never ran its own analysis because the source
was already done) fell back to a single-channel mono waveform while its
siblings of the same source rendered full stereo. The waveform pyramid is
content-addressed per source, so all clips of a media file can share it.

- enrichClipsWithSourceWaveformRef (new util): when a clip has no waveform
  pyramid ref, borrow the shared source id from the media file or any sibling
  clip of the same source. Render-time, non-destructive; fixes existing and
  future clips. Feeds the warmup, main-thread painter, and worker resource
  builder alike (TimelineClipCanvas enriches once at the top).
- Document the per-source pyramid model + "per-clip refs are a cache, not the
  source of truth" in docs/Features/Timeline.md (Audio).

Keep TimelineClipCanvas under its 435-line architecture budget by extracting
cohesive helpers (the host was at the ceiling): clip geometry
(resolveClipGeometry/createWorkerDrawableClips) and viewport sizing
(useTimelineClipCanvasViewport) move to focused modules.

Worker runtime test mocks prefersSoftwareTimelineCanvas to non-Linux so the
worker path is exercised (the Linux software fallback is intentional in prod).

Checks: npm run build + lint (0 errors) + test (4183 passed) all green.
