+++ docs/Features/Timeline.md
@@ -80,6 +80,7 @@ getTrackChildren()  // Query child tracks
 - Simple detailed-waveform updates for those non-destructive region edits derive from the source waveform pyramid when possible, and region gain/silence preview is applied directly to the visible waveform columns while dragging. The lane keeps high-resolution detail without waiting for a full processed-audio render, and derivable processed-waveform refreshes run as background cache updates without showing the red waveform-generation border or progress bar. Scrub and playback preview also follow simple region gain/silence changes instead of always using the raw source level.
 - Audio edit-stack chips appear on edited audio clips for quick state, while the selected clip Properties panel exposes an `Audio Edits` tab for inspection, bypass, removal, clear, bake, unbake, and bake history.
 - Bake renders active region edits into a new WAV media source in the Media Panel `Baked Audio` folder, stores the project-local file under `Raw/Baked Audio/`, and resets the clip edit stack. Unbake restores the latest reversible bake to the original source media and the pre-bake region edit stack when that source media is still present in the project.
+- **Waveform pyramid resolution:** the displayed waveform comes from a pyramid that is content-addressed by the **source** (one pyramid per media file, shared by all its clips). The reference is stored per clip (`audioState.sourceAnalysisRefs.waveformPyramidId`) and is only back-filled onto the clip that triggered analysis (`clipWaveformAnalysisActions.ts`). A clip created/rebuilt before the source finished analysing can therefore end up with an empty ref and fall back to a **single-channel (mono)** render while its siblings render full stereo. `TimelineClipCanvas` guards against this by back-filling a missing ref at render time from the media file or any sibling clip of the same source (`enrichClipsWithSourceWaveformRef`). Per-clip refs are a cache, not the source of truth — never assume a clip without one has no waveform.
 
 ### Text
 - Created through the timeline text slice.
